add support for eida50 view render
