+++ test/test_selectors.py
@@ -1,5 +1,6 @@
 import pytest
 import datetime as dt
+import numpy as np
 from forest import selectors, db
 
 
@@ -12,9 +13,13 @@ from forest import selectors, db
     ("variable", db.State(variable="mslp"), "mslp"),
     ("initial_time", {}, None),
     ("initial_time", {"initial_time": "2019-01-01 00:00:00"}, dt.datetime(2019, 1, 1)),
+    ("initial_time", {"initial_time": "2019-01-01T00:00:00"}, dt.datetime(2019, 1, 1)),
+    ("initial_time", {"initial_time": np.datetime64("2019-01-01 00:00:00", "s")}, dt.datetime(2019, 1, 1)),
     ("initial_time", db.State(initial_time="2019-01-01 00:00:00"), dt.datetime(2019, 1, 1)),
     ("valid_time", {}, None),
     ("valid_time", {"valid_time": "2019-01-01 00:00:00"}, dt.datetime(2019, 1, 1)),
+    ("valid_time", {"valid_time": "2019-01-01T00:00:00"}, dt.datetime(2019, 1, 1)),
+    ("valid_time", {"valid_time": np.datetime64("2019-01-01 00:00:00", "s")}, dt.datetime(2019, 1, 1)),
     ("valid_time", db.State(valid_time="2019-01-01 00:00:00"), dt.datetime(2019, 1, 1)),
     ("pressures", {}, None),
     ("pressures", {"pressure": 1000.}, None),
